% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_disk.R
\name{Qfit_d}
\alias{Qfit_d}
\title{Qfit_d}
\usage{
Qfit_d(k, time, h0 = 0.225, rb = 0.05, tb = 0.01)
}
\arguments{
\item{k}{k}

\item{time}{time}

\item{h0}{h0}

\item{rb}{rb}

\item{tb}{tb}
}
\description{
Returns Qfit
}
\author{
Camille Heylen
}
