% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_disk.R
\name{RMSE_Qfit_d}
\alias{RMSE_Qfit_d}
\title{Qfit_d}
\usage{
RMSE_Qfit_d(k, tb = 0.01, rb = 0.05, h0 = 0.225, data)
}
\arguments{
\item{k}{k}

\item{tb}{tb}

\item{rb}{rb}

\item{h0}{h0}

\item{data}{data}
}
\description{
Returns RMSE
}
\author{
RMSE_Qfit_d Heylen
}
