% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_disk.R
\name{RMSE_Qfit_d_all}
\alias{RMSE_Qfit_d_all}
\alias{RMSE_Qfit_d_allparams}
\title{RMSE_Qfit_d_all}
\usage{
RMSE_Qfit_d_all(k, tb = 0.01, rb = 0.05, h0 = 0.225, data)

RMSE_Qfit_d_allparams(param, data)
}
\arguments{
\item{k}{k}

\item{tb}{tb}

\item{rb}{rb}

\item{h0}{h0}

\item{data}{All data}

\item{k}{k}

\item{tb}{tb}

\item{rb}{rb}

\item{h0}{h0}

\item{data}{All data}
}
\description{
Returns sum of RMSE

Returns sum of RMSE
}
\author{
Camille Heylen

Camille Heylen
}
