% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_disk.R
\name{RMSE_Vfit_d_allparams}
\alias{RMSE_Vfit_d_allparams}
\title{RMSE_Vfit_d_allparams}
\usage{
RMSE_Vfit_d_allparams(param, data)
}
\arguments{
\item{param}{k h0 rb tb}

\item{data}{All data}
}
\description{
Returns sum of RMSE
}
\author{
Camille Heylen
}
