% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filter.R
\name{Vfit_f}
\alias{Vfit_f}
\alias{Vfit_F}
\title{Vfit_f}
\usage{
Vfit_F(k, time, h0 = 0.225, rb = 0.0975, tb = 0.015, rt = 0.135,
  tw = 0.0115, L = 0.235)
}
\arguments{
\item{k}{k}

\item{time}{time}

\item{h0}{h0}

\item{rb}{rb}

\item{tb}{tb}

\item{rt}{rt}

\item{tw}{tw}

\item{L}{L}
}
\description{
Returns Qfit
}
\author{
Camille Heylen
}
